Renamed MappingCreationModal to MappingNameModal since it includes both creation and renaming of mappings now.
